+++ python/packages/azurefunctions/agent_framework_azurefunctions/_feature_usage.py
@@ -1,9 +1,0 @@
-# Copyright (c) Microsoft. All rights reserved.
-
-from enum import IntEnum
-
-
-class FeatureIndex(IntEnum):
-    """Azure Functions-owned feature-usage indexes."""
-
-    AZUREFUNCTIONS = 78