+++ python/packages/durabletask/agent_framework_durabletask/_feature_usage.py
@@ -1,9 +1,0 @@
-# Copyright (c) Microsoft. All rights reserved.
-
-from enum import IntEnum
-
-
-class FeatureIndex(IntEnum):
-    """Durable Task-owned feature-usage indexes."""
-
-    DURABLETASK = 77